+++ src/main/java/com/soft/library/DataBase/dao/ElementDao.java
@@ -12,7 +12,7 @@ import java.util.List;
 public interface ElementDao <E> {
     public void addElement(E element);
     public void updateElement(E element);
-    public E getElementByID(Long elementId);
+    public E getElementByID(Integer elementId);
     public List<E> getAllElements();
     public void deleteElement(E element);
